day 11 part 2

+++ day11/src/main/kotlin/common/InputParser.kt
@@ -2,6 +2,6 @@ package common
 
 import java.io.File
 
-fun getInitialStones(path: String): List<Int> {
-	return File("src/main/resources/$path").readLines()[0].split(" ").map { it.toInt() }
+fun getInitialStones(path: String): List<Long> {
+	return File("src/main/resources/$path").readLines()[0].split(" ").map { it.toLong() }
 }
